Add Recommendations reworked
add_recommendations connects to the database. It presents the data that
it gets from imdb.php.

imdb.php requests data from imdb.com. It filters out the data that does
not meet the requirements (age, score).

engines.php facilitats the call from add_recommendations.php to imdb.php

(bug fix)
The movie that we are getting recommendations for is no longer included
in the result.