php7: remove break that is not part of a loop or switch context
This break is neither part of a loop nor of a switch context and causes
PHP7 to abort with a fatal error. As this break is not reachable due to
the return in the statement before this was not triggered with older
versions but with PHP7s AST this is detected.