Fixed problem with missing: cover image, actor image, countries and languages.
I found on http://stackoverflow.com/questions/31856365/amazon-ecs-api-to-return-secure-images-url
that there is a link between the url in IMDB
https://images-na.ssl-images-amazon.com
and 
http://ecx.images-amazon.com

Also it seams that the last part of the image url is some sort of function that scales the image. And by removing the part between @@ and .jpg we get the full size image.

Also the urls for languages and countries has changed so I have fixed those.