Added missing function, fixes issue #64
I've retrieved the function from CVS, it was removed in 1.11:

----------------------------
revision 1.11
date: 2010/02/14 11:46:21;  author: andig2;  state: Exp;  lines: +1 -55
removed cache analysis functionality
----------------------------

I didn't put it back in setup.core.php since this is the only place it's being used.
Might be usefull for core/cache, but that's something I'm not going to decide right now.
Going for the Fix First Refactor Later option ;-)