Poc for default owner for browsing
Little poc for a default owner for browsing problem as described in #3
This is an issue when you have 'guest'  users, when they log in they first see blank page and always have to select <any> owner from filter before viewing anything. This code fixes that in when a user has zero entries themselves the default owner filter will be set to lang_filter_any instead of their username. This works without bc breaks.